Testing for Map Container
Testing for the following scenarios:
When destination coordinates are present, the component should render with the current location marker and routing machine.
When destination coordinates are not present, the component should render with the current location marker only.
We also use the jest.mock function to mock the Routes component that is imported in the MapCont component. In this case, we are returning a simple div element with a data-testid attribute that we can use to verify that the Routes component is being rendered correctly.